feat(agent): restore gated runtime ownership
Port the scoped runtime split and lifecycle ownership from closed PR #16286 onto current main while preserving subsequent agent and CRDT behavior.

Source-Commit: 770353d9df6e2f252fe74b0dbfeea7d78ffb26a9
Source-Commit: 445931d16853ca34f4b5bb7f66878405b83031f9
