+++ src/workbench/extensions/agent/services/agent/agentRestClient.ts
@@ -55,6 +55,10 @@ export interface DraftSnapshot {
   version?: number
 }
 
+export interface AgentRestClientOptions {
+  signal?: () => AbortSignal | undefined
+}
+
 export interface PostMessageInput {
   content: string
   workflowId?: string
@@ -78,7 +82,7 @@ function isIngestErrorBody(body: unknown): body is IngestErrorBody {
   )
 }
 
-export function createAgentRestClient() {
+export function createAgentRestClient(options: AgentRestClientOptions = {}) {
   async function toApiError(response: Response): Promise<AgentApiError> {
     const text = await response.text()
     let body: unknown
@@ -101,7 +105,11 @@ export function createAgentRestClient() {
     init: RequestInit,
     schema: z.ZodType<T>
   ): Promise<T> {
-    const response = await api.fetchApi(route, init)
+    const signal = options.signal?.()
+    const response = await api.fetchApi(
+      route,
+      signal === undefined ? init : { ...init, signal }
+    )
     if (!response.ok) throw await toApiError(response)
     return schema.parse(await response.json())
   }
